+++ include/roydeque.h
@@ -19,7 +19,7 @@ typedef struct RoyDeque_ RoyDeque;
 // Returns a pointer to a newly build RoyDeque. 
 RoyDeque * roy_deque_new(size_t element_size);
 
-// Deallocates all the memory allocated.
+// De-allocates all the memory allocated.
 // (Always call this function after the work is done by the given 'deque', or memory leak will occur.)
 void roy_deque_delete(RoyDeque * deque);
 
@@ -28,7 +28,7 @@ void roy_deque_delete(RoyDeque * deque);
 // Returns an iterator to 'position' in 'deque' where the element takes place.
 // (If 'position' is in small half of the deque, iteration will start at head, and vice versa.)
 // (Returns NULL if position is out of range.)
-void * roy_deque_pointer(RoyDeque * deque, int position);
+void * roy_deque_pointer(RoyDeque * deque, size_t position);
 
 // Returns an iterator to the element at front end.
 void * roy_deque_front(RoyDeque * deque);
@@ -39,7 +39,7 @@ void * roy_deque_back(RoyDeque * deque);
 // Returns an iterator to 'position' in 'deque' where the element takes place.
 // (If 'position' is in small half of the deque, iteration will start at head, and vice versa.)
 // (Returns NULL if position is out of range.)
-const void * roy_deque_const_pointer(const RoyDeque * deque, int position);
+const void * roy_deque_const_pointer(const RoyDeque * deque, size_t position);
 
 // Returns an const iterator to the element at front end.
 const void * roy_deque_const_front(const RoyDeque * deque);
@@ -49,7 +49,7 @@ const void * roy_deque_const_back(const RoyDeque * deque);
 
 // Returns a copy of the element at 'position'. (With boundary check)
 // (The behavior is undefined if 'dest' is uninitialized.)
-void * roy_deque_element(void * dest, const RoyDeque * deque, int position);
+void * roy_deque_element(void * dest, const RoyDeque * deque, size_t position);
 
 // Returns a typed pointer to the element at 'position', NULL if position exceeds.
 #define roy_deque_at(deque, element_type, position) ((element_type *)(roy_deque_pointer((deque), (position))))
@@ -65,7 +65,7 @@ bool roy_deque_empty(const RoyDeque * deque);
 /* MODIFIERS */
 
 // Add an element named 'data' into 'deque' at 'position'.
-RoyDeque * roy_deque_insert(RoyDeque * deque, int position, const void * data);
+RoyDeque * roy_deque_insert(RoyDeque * deque, size_t position, const void * data);
 
 // Adds an element named 'data' at the beginning of 'deque'.
 RoyDeque * roy_deque_push_front(RoyDeque * deque, const void * data);
@@ -74,7 +74,7 @@ RoyDeque * roy_deque_push_front(RoyDeque * deque, const void * data);
 RoyDeque * roy_deque_push_back(RoyDeque * deque, const void * data);
 
 // Removes an element from 'deque' at 'position'.
-RoyDeque * roy_deque_erase(RoyDeque * deque, int position);
+RoyDeque * roy_deque_erase(RoyDeque * deque, size_t position);
 
 // Removes the first element from 'deque'.
 RoyDeque * roy_deque_pop_front(RoyDeque * deque);
